+++ tools/vscode-ext/src/extension.ts
@@ -1,12 +1,11 @@
+import * as fs from 'fs';
 import * as path from 'path';
-import { window } from 'vscode';
-import { workspace, ExtensionContext } from 'vscode';
+import { ExtensionContext } from 'vscode';
 
 import {
 	LanguageClient,
 	LanguageClientOptions,
-	ServerOptions,
-	TransportKind
+	ServerOptions
 } from 'vscode-languageclient/node';
 
 export function activate(context: ExtensionContext) {
@@ -17,11 +16,13 @@ function tryStartLanguageServer(context: ExtensionContext) {
 	//for future, possibly may want version checking
 
 	console.log("Before setup");
-	const serverPath = path.join(__dirname, '../../../target/release/conjure-oxide');
-    
-    let serveroptions: ServerOptions = {
-        command: serverPath, args: ["server-lsp"]
-    }
+	const localServerPath = path.join(__dirname, '../../../target/release/conjure-oxide');
+	const command = fs.existsSync(localServerPath) ? localServerPath : "conjure-oxide";
+
+	let serveroptions: ServerOptions = {
+		command,
+		args: ["server-lsp"]
+	}
 
 	let clientoptions: LanguageClientOptions = {
 		documentSelector: [{ scheme: 'file', language: 'essence' }]